+++ RichMessageKit/Views/SentButtonsCell.swift
@@ -9,6 +9,7 @@ import Foundation
 
 public class SentButtonsCell: UITableViewCell {
     // MARK: - Public properties
+
     public struct Config {
         public static var buttonTopPadding: CGFloat = 4
         public static var padding = Padding(left: 60, right: 10, top: 10, bottom: 10)
@@ -18,7 +19,7 @@ public class SentButtonsCell: UITableViewCell {
 
     // MARK: - Fileprivate properties
 
-    fileprivate lazy var buttons = SuggestedReplyView(maxWidth: Config.buttonWidth)
+    fileprivate lazy var buttons = SuggestedReplyView()
     fileprivate lazy var messageView = SentMessageView(
         frame: .zero,
         padding: messageViewPadding,
@@ -58,7 +59,7 @@ public class SentButtonsCell: UITableViewCell {
             maxWidth: Config.maxWidth,
             padding: messageViewPadding
         )
-        buttons.update(model: model)
+        buttons.update(model: model, maxWidth: Config.buttonWidth)
     }
 
     /// It is used to get exact height of `SentButtonsCell` using messageModel, width and padding
@@ -89,7 +90,6 @@ public class SentButtonsCell: UITableViewCell {
             buttons.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -Config.padding.right),
             buttons.leadingAnchor.constraint(equalTo: leadingAnchor),
             buttons.bottomAnchor.constraint(equalTo: self.bottomAnchor, constant: -1 * Config.padding.bottom),
-            ])
+        ])
     }
-
 }